fix: perform getCatalog requests in batches
